Change up multi-bit picross puzzles.
Instead of a total, the additional hints show how many of each shade there are, and the number of changes in shading when reading the nonzero cells from top/bottom left/right. After playtesting this seems to lead to much less trial and error (though still multiple solutions can exist).

Change up the protobuf format to reflect new options, it is not compatible with previous versions.